+++ src/main/scala/firrtl/passes/memlib/AnnotateMemMacros.scala
@@ -1,6 +1,7 @@
 // See LICENSE for license details.
 
 package firrtl.passes
+package memlib
 
 import firrtl._
 import firrtl.ir._
@@ -9,19 +10,14 @@ import firrtl.Mappers._
 import WrappedExpression.weq
 import MemPortUtils.memPortField
 import AnalysisUtils._
-
-case class AppendableInfo(fields: Map[String, Any]) extends Info {
-  def append(a: Map[String, Any]) = this.copy(fields = fields ++ a)
-  def append(a: (String, Any)): AppendableInfo = append(Map(a))
-  def get(f: String) = fields.get(f)
-  override def equals(b: Any) = b match {
-    case i: AppendableInfo => fields - "info" == i.fields - "info"
-    case _ => false
-  }
-}
+import AppendableUtils._
 
 object AnalysisUtils {
   type Connects = collection.mutable.HashMap[String, Expression]
+
+  /** Builds a map from named component to assigned value
+    * Named components are serialized LHS of connections, nodes, invalids
+    */
   def getConnects(m: DefModule): Connects = {
     def getConnects(connects: Connects)(s: Statement): Statement = {
       s match {
@@ -29,6 +25,8 @@ object AnalysisUtils {
           connects(loc.serialize) = expr
         case DefNode(_, name, value) =>
           connects(name) = value
+        case IsInvalid(_, value) =>
+          connects(value.serialize) = WInvalid
         case _ => // do nothing
       }
       s map getConnects(connects)
@@ -38,25 +36,21 @@ object AnalysisUtils {
     connects
   }
 
-  // takes in a list of node-to-node connections in a given module and looks to find the origin of the LHS.
-  // if the source is a trivial primop/mux, etc. that has yet to be optimized via constant propagation,
-  // the function will try to search backwards past the primop/mux. 
-  // use case: compare if two nodes have the same origin
-  // limitation: only works in a module (stops @ module inputs)
-  // TODO: more thorough (i.e. a + 0 = a)
-  def getConnectOrigin(connects: Connects)(node: String): Expression =
-    connects get node match {
-      case None => EmptyExpression
-      case Some(e) => getOrigin(connects, e)
-    }
-  def getConnectOrigin(connects: Connects, e: Expression): Expression =
-    getConnectOrigin(connects)(e.serialize)
-
-  private def getOrigin(connects: Connects, e: Expression): Expression = e match { 
+  /** Find a connection LHS's origin from a module's list of node-to-node connections
+    *   regardless of whether constant propagation has been run.
+    * Will search past trivial primop/mux's which do not affect its origin.
+    * Limitations: 
+    *   - Only works in a module (stops @ module inputs)
+    *   - Only does trivial primop/mux's (is not complete)
+    * TODO(shunshou): implement more equivalence cases (i.e. a + 0 = a)
+    */
+  def getOrigin(connects: Connects, s: String): Expression =
+    getOrigin(connects)(WRef(s, UnknownType, ExpKind, UNKNOWNGENDER))
+  def getOrigin(connects: Connects)(e: Expression): Expression = e match { 
     case Mux(cond, tv, fv, _) =>
-      val fvOrigin = getOrigin(connects, fv)
-      val tvOrigin = getOrigin(connects, tv)
-      val condOrigin = getOrigin(connects, cond)
+      val fvOrigin = getOrigin(connects)(fv)
+      val tvOrigin = getOrigin(connects)(tv)
+      val condOrigin = getOrigin(connects)(cond)
       if (weq(tvOrigin, one) && weq(fvOrigin, zero)) condOrigin
       else if (weq(condOrigin, one)) tvOrigin
       else if (weq(condOrigin, zero)) fvOrigin
@@ -69,54 +63,55 @@ object AnalysisUtils {
       val extractionWidth = (msb - lsb) + 1
       val nodeWidth = bitWidth(args.head.tpe)
       // if you're extracting the full bitwidth, then keep searching for origin
-      if (nodeWidth == extractionWidth) getOrigin(connects, args.head) else e
+      if (nodeWidth == extractionWidth) getOrigin(connects)(args.head) else e
     case DoPrim((PrimOps.AsUInt | PrimOps.AsSInt | PrimOps.AsClock), args, _, _) => 
-      getOrigin(connects, args.head)
-    // Todo: It's not clear it's ok to call remove validifs before mem passes...
-    case ValidIf(cond, value, ClockType) => getOrigin(connects, value)
+      getOrigin(connects)(args.head)
+    case ValidIf(cond, value, ClockType) => getOrigin(connects)(value)
     // note: this should stop on a reg, but will stack overflow for combinational loops (not allowed)
     case _: WRef | _: WSubField | _: WSubIndex | _: WSubAccess if kind(e) != RegKind =>
        connects get e.serialize match {
-         case Some(ex) => getOrigin(connects, ex)
+         case Some(ex) => getOrigin(connects)(ex)
          case None => e
        }
     case _ => e
   }
 
-  def appendInfo[T <: Info](info: T, add: Map[String, Any]) = info match {
-    case i: AppendableInfo => i.append(add)
-    case _ => AppendableInfo(fields = add + ("info" -> info))
-  }
-  def appendInfo[T <: Info](info: T, add: (String, Any)): AppendableInfo = appendInfo(info, Map(add))
-  def getInfo[T <: Info](info: T, k: String) = info match {
-    case i: AppendableInfo => i.get(k)
-    case _ => None
-  }
-  def containsInfo[T <: Info](info: T, k: String) = info match {
-    case i: AppendableInfo => i.fields.contains(k)
-    case _ => false
-  }
-
-  // memories equivalent as long as all fields (except name) are the same
+  /** Checks whether the two memories are equivalent in all respects except name
+    */
   def eqMems(a: DefMemory, b: DefMemory) = a == b.copy(name = a.name) 
 }
 
+/** Annotates sequential memories that are candidates for macro replacement.
+  * Annotations are held in DefMemory's info field as AppendableInfo
+  * Requirements for macro replacement:
+  *   - read latency and write latency of one
+  *   - only one readwrite port or write port
+  *   - zero or one read port
+  * In addition, determines if a write mask is needed (wmode/en and wmask are equivalent).
+  * Annotations:
+  *   - "useMacro" -- is true if a valid candidate
+  *   - "maskGran" -- (dataType size) / (number of mask bits)
+  *      - i.e. 1 if bitmask, 8 if bytemask, absent for no mask
+  * TODO(shunshou): Add floorplan info?
+  */
 object AnnotateMemMacros extends Pass {
-  def name = "Analyze sequential memories and tag with info for future passes(useMacro, maskGran)"
+  def name = "Annotate Memory Macros"
 
-  // returns # of mask bits if used
+  /** Returns the number of mask bits, if used
+    */
   def getMaskBits(connects: Connects, wen: Expression, wmask: Expression): Option[Int] = {
-    val wenOrigin = getConnectOrigin(connects, wen)
+    val wenOrigin = getOrigin(connects)(wen)
     val wmaskOrigin = connects.keys filter
-      (_ startsWith wmask.serialize) map getConnectOrigin(connects)
+      (_ startsWith wmask.serialize) map {s: String => getOrigin(connects)(WRef(s, UnknownType, ExpKind, UNKNOWNGENDER))}
     // all wmask bits are equal to wmode/wen or all wmask bits = 1(for redundancy checking)
     val redundantMask = wmaskOrigin forall (x => weq(x, wenOrigin) || weq(x, one))
     if (redundantMask) None else Some(wmaskOrigin.size)
   }
 
+  /** Only annotate memories that are candidates for memory macro replacements
+    * i.e. rw, w + r (read, write 1 cycle delay)
+    */
   def updateStmts(connects: Connects)(s: Statement): Statement = s match {
-    // only annotate memories that are candidates for memory macro replacements
-    // i.e. rw, w + r (read, write 1 cycle delay)
     case m: DefMemory if m.readLatency == 1 && m.writeLatency == 1 &&
         (m.writers.length + m.readwriters.length) == 1 && m.readers.length <= 1 =>
       val dataBits = bitWidth(m.dataType)
@@ -136,8 +131,6 @@ object AnnotateMemMacros extends Pass {
   }
 
   def annotateModMems(m: DefModule) = m map updateStmts(getConnects(m))
-
   def run(c: Circuit) = c copy (modules = c.modules map annotateModMems)
 }
 
-// TODO: Add floorplan info?
