+++ src/main/scala/firrtl/passes/memlib/MemTransformUtils.scala
@@ -1,0 +1,47 @@
+package firrtl.passes
+package memlib
+
+import firrtl._
+import firrtl.ir._
+import firrtl.Utils._
+import firrtl.Mappers._
+import AnalysisUtils._
+import MemPortUtils._
+import AppendableUtils._
+
+object MemTransformUtils {
+
+  /** Replaces references to old memory port names with new memory port names
+   */
+  def updateStmtRefs(repl: MemPortMap)(s: Statement): Statement = {
+    //TODO(izraelevitz): check speed
+    def updateRef(e: Expression): Expression = {
+      val ex = e map updateRef
+      repl getOrElse (ex.serialize, ex)
+    }
+
+    def hasEmptyExpr(stmt: Statement): Boolean = {
+      var foundEmpty = false
+      def testEmptyExpr(e: Expression): Expression = {
+        e match {
+          case EmptyExpression => foundEmpty = true
+          case _ =>
+        }
+        e map testEmptyExpr // map must return; no foreach
+      }
+      stmt map testEmptyExpr
+      foundEmpty
+    }
+
+    def updateStmtRefs(s: Statement): Statement =
+      s map updateStmtRefs map updateRef match {
+        case c: Connect if hasEmptyExpr(c) => EmptyStmt
+        case s => s
+      }
+
+    updateStmtRefs(s)
+  }
+
+}
+
+// vim: set ts=4 sw=4 et:
